scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.2.0",
]
